Split MVT to GeoJSON translation into separate files and add GeoEngine Configuration
The GeoEngine configuration contains the grid size and zoom levels along with the
servers. This used to be hidden away in TileGridUtils.kt which wasn't the best
place for it.
